+++ api/typescript/src/core/kurtosis_core_rpc_api_bindings/api_container_service_pb.js
@@ -10172,7 +10172,8 @@ proto.api_container_api.HttpRequestFactRecipe.toObject = function(includeInstanc
     endpoint: jspb.Message.getFieldWithDefault(msg, 2, ""),
     method: jspb.Message.getFieldWithDefault(msg, 3, 0),
     contentType: jspb.Message.getFieldWithDefault(msg, 4, ""),
-    body: jspb.Message.getFieldWithDefault(msg, 5, "")
+    body: jspb.Message.getFieldWithDefault(msg, 5, ""),
+    fieldExtractor: jspb.Message.getFieldWithDefault(msg, 6, "")
   };
 
   if (includeInstance) {
@@ -10228,6 +10229,10 @@ proto.api_container_api.HttpRequestFactRecipe.deserializeBinaryFromReader = func
     case 5:
       var value = /** @type {string} */ (reader.readString());
       msg.setBody(value);
+      break;
+    case 6:
+      var value = /** @type {string} */ (reader.readString());
+      msg.setFieldExtractor(value);
       break;
     default:
       reader.skipField();
@@ -10290,6 +10295,13 @@ proto.api_container_api.HttpRequestFactRecipe.serializeBinaryToWriter = function
   if (f.length > 0) {
     writer.writeString(
       5,
+      f
+    );
+  }
+  f = /** @type {string} */ (jspb.Message.getField(message, 6));
+  if (f != null) {
+    writer.writeString(
+      6,
       f
     );
   }
@@ -10383,6 +10395,42 @@ proto.api_container_api.HttpRequestFactRecipe.prototype.getBody = function() {
  */
 proto.api_container_api.HttpRequestFactRecipe.prototype.setBody = function(value) {
   return jspb.Message.setProto3StringField(this, 5, value);
+};
+
+
+/**
+ * optional string field_extractor = 6;
+ * @return {string}
+ */
+proto.api_container_api.HttpRequestFactRecipe.prototype.getFieldExtractor = function() {
+  return /** @type {string} */ (jspb.Message.getFieldWithDefault(this, 6, ""));
+};
+
+
+/**
+ * @param {string} value
+ * @return {!proto.api_container_api.HttpRequestFactRecipe} returns this
+ */
+proto.api_container_api.HttpRequestFactRecipe.prototype.setFieldExtractor = function(value) {
+  return jspb.Message.setField(this, 6, value);
+};
+
+
+/**
+ * Clears the field making it undefined.
+ * @return {!proto.api_container_api.HttpRequestFactRecipe} returns this
+ */
+proto.api_container_api.HttpRequestFactRecipe.prototype.clearFieldExtractor = function() {
+  return jspb.Message.setField(this, 6, undefined);
+};
+
+
+/**
+ * Returns whether this field is set.
+ * @return {boolean}
+ */
+proto.api_container_api.HttpRequestFactRecipe.prototype.hasFieldExtractor = function() {
+  return jspb.Message.getField(this, 6) != null;
 };